acrtualizacion 13 sep

+++ database/migrations/2022_05_01_133619_create_conf_aportes_table.php
@@ -16,13 +16,14 @@ class CreateConfAportesTable extends Migration
         Schema::create('conf_aportes', function (Blueprint $table) {
             $table->id();
             $table->string('tipo_aporte',100);
-            $table->integer('rango_inicial');
-            $table->integer('rango_final');
-            $table->float('porcentaje_aporte',9,2);
+            $table->decimal('rango_inicial',9,2);
+            $table->decimal('rango_final',9,2);
+            $table->decimal('porcentaje_aporte',9,2);
             $table->enum('estado', ['HABILITADO', 'INHABILITADO'])->default('HABILITADO');
             $table->timestamps();
         });
     }
+
 
     /**
      * Reverse the migrations.
